Move impl of Verifier::verify into the trait
